+++ lib/utils/analytics.dart
@@ -5,8 +5,12 @@ import 'package:myagenda/models/analytics.dart';
 import 'package:myagenda/utils/preferences.dart';
 
 class AnalyticsProvider extends InheritedWidget {
-  AnalyticsProvider(this.analytics, this.observer, {Key key, this.child})
-      : super(key: key, child: child);
+  AnalyticsProvider(
+    this.analytics,
+    this.observer, {
+    Key key,
+    this.child,
+  }) : super(key: key, child: child);
 
   final FirebaseAnalytics analytics;
   final FirebaseAnalyticsObserver observer;
@@ -24,7 +28,7 @@ class AnalyticsProvider extends InheritedWidget {
       AnalyticsValue.groupKeys:
           prefs.urlIcs != null ? "Ical File" : prefs.groupKeys?.join(',') ?? "",
       AnalyticsValue.university:
-          prefs.urlIcs ?? prefs.university?.name ?? "Unknown",
+          prefs.urlIcs ?? prefs.university?.university ?? "Unknown",
     });
   }