+++ lib/utils/login/login_cas.dart
@@ -1,45 +1,40 @@
 import 'dart:async';
 import 'package:html/parser.dart' show parse;
+import 'package:myagenda/models/preferences/university.dart';
 import 'package:myagenda/utils/http/http_request.dart';
 import 'package:myagenda/utils/login/login_base.dart';
 
 class LoginCAS extends LoginProcess {
-  LoginCAS(loginUrl, username, password) : super(loginUrl, username, password);
+  LoginCAS(University university, String username, String password)
+      : super(university, username, password);
 
   Future<LoginResult> login() async {
-    final response = await HttpRequest.get(loginUrl);
+    final response = await HttpRequest.get(university.loginUrl);
 
     if (!response.isSuccess) return LoginResult(LoginResultType.NETWORK_ERROR);
 
     final document = parse(response.httpResponse.body);
 
-    // Extract lt value from HTML
-    final ltInput = document.querySelector('input[name="lt"]');
-    String lt = "";
-    if (ltInput?.attributes?.containsKey("value") ?? false)
-      lt = ltInput.attributes['value'];
-
-    final executionInput = document.querySelector('input[name="execution"]');
-    String execution;
-    if (executionInput?.attributes?.containsKey("value") ?? false)
-      execution = executionInput.attributes['value'];
-
     // POST data
-    Map<String, String> postParams = {
-      "_eventId": "submit",
-      "lt": lt,
-      "submit": "LOGIN",
-      "username": username,
-      "password": password,
-    };
-    if (execution != null) postParams['execution'] = execution;
+    Map<String, String> postParams = {};
+    postParams[university.credentialFields.username] = username;
+    postParams[university.credentialFields.password] = password;
+
+    // Extract fields value from HTML
+    university.loginFields.forEach((field) {
+      final fieldElem = document.querySelector('input[name="' + field + '"]');
+      String fieldValue = "";
+      if (fieldElem?.attributes?.containsKey('value') ?? false)
+        fieldValue = fieldElem.attributes['value'];
+      postParams[field] = fieldValue;
+    });
 
     // Get JSESSIONID from previous request header
     final cookie = response.httpResponse.headers["set-cookie"];
 
     // Second request, with all necessary data
     final loginResponse = await HttpRequest.post(
-      loginUrl,
+      university.loginUrl,
       body: postParams,
       headers: {"cookie": cookie},
     );
@@ -48,16 +43,17 @@ class LoginCAS extends LoginProcess {
       return LoginResult(LoginResultType.NETWORK_ERROR);
 
     final loginDocument = parse(loginResponse.httpResponse.body);
-    final errorElement = loginDocument.querySelector(".errors");
-    final successElement = loginDocument.querySelector(".success");
+    final errorElement = loginDocument.querySelector(
+      university.statusTags.error,
+    );
+    final successElement = loginDocument.querySelector(
+      university.statusTags.success,
+    );
 
-    if (errorElement != null && successElement == null) {
+    if (errorElement != null) {
       // Display error to user
-      String error = errorElement.innerHtml;
-      error = error.replaceAll("<h2>", "").replaceAll("</h2>", "").trim();
-
-      return LoginResult(LoginResultType.LOGIN_FAIL, error);
-    } else if (successElement != null && errorElement == null) {
+      return LoginResult(LoginResultType.LOGIN_FAIL, errorElement.innerHtml);
+    } else if (successElement != null) {
       return LoginResult(LoginResultType.LOGIN_SUCCESS);
     } else {
       return LoginResult(LoginResultType.UNKNOWN_ERROR);