+++ lib/widgets/settings/list_tile_number.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:myagenda/keys/string_key.dart';
 import 'package:myagenda/screens/base_state.dart';
+import 'package:myagenda/utils/translations.dart';
 import 'package:myagenda/widgets/settings/list_tile_title.dart';
 import 'package:myagenda/widgets/ui/dialog/dialog_predefined.dart';
 import 'package:myagenda/widgets/ui/number_selector.dart';
@@ -77,8 +78,8 @@ class _ListTileNumberState extends BaseState<ListTileNumber> {
           initialValue: _inputValue,
           onChanged: _onInputChange,
         ),
-        translation(StrKey.SUBMIT),
-        translation(StrKey.CANCEL),
+        translations.text(StrKey.SUBMIT),
+        translations.text(StrKey.CANCEL),
         true,
         const EdgeInsets.all(0.0));