+++ lib/widgets/ui/screen_message/empty_day.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_i18n/flutter_i18n.dart';
 import 'package:myagenda/keys/string_key.dart';
+import 'package:myagenda/utils/translations.dart';
 
 class EmptyDay extends StatelessWidget {
   final EdgeInsets padding;
@@ -13,7 +13,7 @@ class EmptyDay extends StatelessWidget {
       padding: padding ??
           const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
       child: Text(
-        FlutterI18n.translate(context, StrKey.NO_EVENTS),
+        translations.text(StrKey.NO_EVENTS),
         style: Theme.of(context).textTheme.subhead,
         textAlign: TextAlign.justify,
       ),